Example Repository to show error in HarpJS
RFC 1738 does not forbid dots in URL's as reserved or special in any way
and therefore harp's routing does not conform to this.

	https://www.ietf.org/rfc/rfc1738.txt

While harp _does_ give clean urls and strips extensions, I was not under
the impression it would not strip more than the last dot and extension
away.